refactor(web): move metrics from shared to infra
metrics is only consumed by infra/run and API routes related to
infrastructure concerns. It belongs in the infra layer, not shared.
