+++ turbo/apps/web/app/api/agent/runs/[id]/telemetry/system-log/__tests__/route.test.ts
@@ -12,7 +12,7 @@ import {
 } from "../../../../../../../../src/__tests__/test-helpers";
 import { mockClerk } from "../../../../../../../../src/__tests__/clerk-mock";
 import { randomUUID } from "crypto";
-import * as metricsModule from "../../../../../../../../src/lib/shared/metrics";
+import * as metricsModule from "../../../../../../../../src/lib/infra/metrics";
 
 const context = testContext();
