+++ turbo/apps/web/app/api/webhooks/agent/telemetry/__tests__/route.test.ts
@@ -11,7 +11,7 @@ import {
   type UserContext,
 } from "../../../../../../src/__tests__/test-helpers";
 import { mockClerk } from "../../../../../../src/__tests__/clerk-mock";
-import * as metricsModule from "../../../../../../src/lib/shared/metrics";
+import * as metricsModule from "../../../../../../src/lib/infra/metrics";
 import * as axiomClient from "../../../../../../src/lib/shared/axiom/client";
 import type { MockInstance } from "vitest";
