+++ turbo/apps/web/src/lib/infra/run/executors/runner-executor.ts
@@ -15,7 +15,7 @@ import { isOfficialRunnerGroup } from "../runner-group";
 import { forbidden } from "../../../shared/errors";
 import { publishJobNotification } from "../../../shared/realtime/client";
 import { logger } from "../../../shared/logger";
-import { recordSandboxOperation } from "../../../shared/metrics";
+import { recordSandboxOperation } from "../../metrics";
 import type { PreparedContext, ExecutorResult } from "./types";
 
 const log = logger("executor:runner");
