+++ turbo/apps/web/src/lib/infra/run/run-service.ts
@@ -28,7 +28,7 @@ import { generateSandboxToken } from "../../auth/sandbox-token";
 import type { ExecutionContext, DispatchTimings, ResumeSession } from "./types";
 import type { ArtifactSnapshot } from "../checkpoint/types";
 import { buildInfraExecutionContext } from "./context/build-context";
-import { recordSandboxOperation } from "../../shared/metrics";
+import { recordSandboxOperation } from "../metrics";
 import { canAccessCompose } from "../agent/compose-access";
 
 import { encryptSecretValue } from "../../shared/crypto/secrets-encryption";
